display statistics and match rules from the D-Bus Stats interface
This requires dbus 1.9.0 for the new GetAllMatchRules call on the Stats
interface and for compiling the Stats interface by default.

The Stats interface is enabled by default on the session bus. On the system
bus, it is not enabled for non-root users by default but it can be enabled as
shown in /usr/share/doc/dbus/examples/example-system-enable-stats.conf

If the feature is not available in D-Bus, D-Feet ignores the error and
continues as before.

https://bugzilla.gnome.org/show_bug.cgi?id=735167