Detect 'pycodestyle', the new name for 'pep8'
The 'pep8' tool has been renamed upstream to 'pycodestyle'. On Debian,
the Python 3 version is packaged and installed as pycodestyle. On
Fedora, it's packaged as python3-pycodestyle and installed as
pycodestyle-3; the Python 2 version is installed as pycodestyle-2 with a
symlink from pycodestyle. Make configure search a variety of such names,
and use the path it finds.